API: move bit_generator and generator to be private
