+++ numpy/random/_bit_generator.pxd
