+++ numpy/random/_mt19937.pyx
@@ -4,7 +4,7 @@ import numpy as np
 cimport numpy as np
 
 from libc.stdint cimport uint32_t, uint64_t
-from .bit_generator cimport BitGenerator, SeedSequence
+from ._bit_generator cimport BitGenerator, SeedSequence
 
 __all__ = ['MT19937']
