+++ numpy/random/_philox.pyx
@@ -10,7 +10,7 @@ cimport numpy as np
 
 from libc.stdint cimport uint32_t, uint64_t
 from ._common cimport uint64_to_double, int_to_array, wrap_int
-from .bit_generator cimport BitGenerator
+from ._bit_generator cimport BitGenerator
 
 __all__ = ['Philox']
