+++ numpy/random/_sfc64.pyx
@@ -3,7 +3,7 @@ cimport numpy as np
 
 from libc.stdint cimport uint32_t, uint64_t
 from ._common cimport uint64_to_double
-from .bit_generator cimport BitGenerator
+from ._bit_generator cimport BitGenerator
 
 __all__ = ['SFC64']
